Fix race condition that needlessly showed the refresh-button
The monitoring support is meant to be something simple and not overcomplicated,
therefore there is a single point-of-entry that marks a window as being edited
from an external tool. While a window contains several notes, saving more than
one of them shifted by 1 second results in false positive. The simple fix to
avoid this race condition is currently to increase the window-updated timeout.

Also in this commit, when the user decides to not reload the notes from the
disk, save the current notes right back to the disk.