gtk3: Implement popup command with Gtk.Application
Instead of depending on oldish X11 code to show/hide the notes, use
Gtk.Application. For GTK3, the binary xfce4-popup-notes is build out of
main-popup.vala, while xfce4-popup-notes.c is kept as-is for GTK2.

Both xfce4-popup-notes and xfce4-notes binaries define a unique ID
org.xfce.Notes and send an activate signal in case a primary instance of
xfce4-notes is already running.