Fix mouse cursor behavior after clicking on a link
When notes window was set to "always on top", the mouse cursor changed
to normal after clicking on a link. But the cursor_over_link value did
not change, so moving the mouse over the same link did not change the
cursor back to hand_cursor.

This behavior was due to the default GtkTextView "state-flags-changed"
signal handler.