Save windows configuration when in system tray mode
This will save the windows configuration when the top-level window goes
into the background. The configuration is written only when it is changed,
so it should not wear out the SSD too much.

According to the documentation, xfce-notes-plugin should save windows
configuration every 10 minutes, as well as automatically save windows
configuration when you logout, reboot or power down.

Therefore, this change only applies to the xfce-notes application.

Related documentation:
https://developer.xfce.org/xfce4-panel/XfcePanelPlugin.html#XfcePanelPlugin-save