Avoid focus loss for note windows
The reason for loss of focus is the code in xfwm4 which is responsible
for finding the next top window when the current window disappears.

The corresponding code can be found in xfwm4/src/focus.c:
clientPassFocus() -> clientGetTopMostFocusable()

This function calls clientGetTopMostFocusable(), which returns
clientPair whose preferred client is not set to one of the xfce4-notes
windows.

This happens because clientSelectMask() returns FALSE when we have
CLIENT_FLAG_SKIP_TASKBAR set.

Let's avoid this behavior by temporarily disabling CLIENT_FLAG_SKIP_TASKBAR
on the window that should catch focus.

Fixes: https://gitlab.xfce.org/panel-plugins/xfce4-notes-plugin/-/issues/56