Use timer instead of idle to handle file events
This allows us to gracefully handle note file editing in vim.

Vim generates a series of actions when a file is saved. Some of these
include renaming the target file, recreating it, and filling it with
content.

To handle this properly, we need to accumulate all file actions in our
queue so that we can optimize it later.