Add support for checkboxes for to-do lists
This doesn't work well with the GtkSourceView Undo/Redo manager, so we
reset it when adding/removing checkboxes to avoid breakage.

When saving a note to a file, checkboxes are surrounded by invisible
spaces, just like text formatting tags. Therefore, they should not
interfere with the user's ability to enter text.

Hotkey used: Ctrl-O

Fixes: https://gitlab.xfce.org/panel-plugins/xfce4-notes-plugin/-/issues/76